requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
